Rename api -> deepwell, web -> framerail, router -> wws.
Since WWS is no longer really a router, the name isn't too relevant.
While we're at it, let's use service names instead of generic names
to avoid this kind of thing in the future.

I implement the changes to the GitHub workflow files in separate
commits, since the diffs git has chosen are quite confusing and
annoying.